Type correction and using own mixin
