+++ .bitmap
@@ -1581,6 +1581,12 @@
         "mainFile": "index.ts",
         "rootDir": "malty/utils/mixins"
     },
+    "utils/space": {
+        "scope": "",
+        "version": "",
+        "mainFile": "index.ts",
+        "rootDir": "malty/utils/space"
+    },
     "utils/test": {
         "scope": "carlsberggroup.malty",
         "version": "0.0.16",
