+++ malty/utils/mixins/index.ts
@@ -9,17 +9,3 @@ export const srOnly = css`
   height: 1px;
   overflow: hidden;
 `;
-
-// TODO: I would like to do something like <T extends SpaceProps>(props: T), but then I get a warning in the style
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const space = ({ m = '', mx = '', my = '', mt = '', mr = '', mb = '', ml = '' }: any) => `
-  margin: ${m};
-  margin-left: ${mx};
-  margin-right: ${mx};
-  margin-top: ${my};
-  margin-bottom: ${my};
-  margin-top: ${mt};
-  margin-right: ${mr};
-  margin-bottom: ${mb};
-  margin-left: ${ml};
-`;
