--- malty/utils/types/index.ts
@@ -1,9 +1,0 @@
-export interface SpaceProps {
-  m?: string;
-  mt?: string;
-  mr?: string;
-  mb?: string;
-  ml?: string;
-  mx?: string;
-  my?: string;
-}
